lib/oelite/meta/sqlite.py: remove unused file
lib/oelite/meta/sqlite.py doesn't seem to be used at all (I can't find
any import of it, and the fact that no .pyc version materialises when
using oe-lite suggests that python never looks at it). It seems to
implement some of the logic which lib/oelite/meta/{dict,meta}.py
provide in pure Python, so I'm guessing it might have been an attempt
at moving more metadata to sql. However, I think that by now we have
consensus that that is the wrong way to go.

Removing this file is an easy first step towards killing remaining uses
of an in-memory sqlite database.