oelite: runq.py, package.py: move recdepend info out from sqlite
While the index on the runq.recdepend table helped quite a lot, the
recdepend info is more naturally kept as a property of the
OElitePackage object itself. This way, there's no need for any
indexing or translating to and from integer ids at all: We just store
and retrieve a set of Python objects (well, the caller expects to
retrieve a list, so we do that for him).

The "Building dependency tree time" drops from ~25s to ~18s for an "oe
bake world". It should also reduce memory usage slightly, since we don't
have to store the package id and deptype once for every row of the
table, along with whatever memory the index consumed.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>