oelite: add show-provided subcommand
It can be frustrating trying to help track down problems which end up being
caused by the other guy using a different autotools version.

So "oe show-provided" does a semi-automatic extraction of the version of
all ASSUME_PROVIDED utils. In many cases, we can just do "$util
--version" and extract the first substring matching
"\b([0-9][0-9.a-zA-Z_-]+)\b" - when necessary, this heuristic can be
overridden, and I've included a few examples.

We can now also attach minimum and maximum versions to the
assume_provided utilities. We're not going to check or enforce these
during normal "oe bake", but it's nice to be able to do "oe
show-provided -c".

A previous version had these minimum and maximum versions hardcoded in
show_provided.py, but it's better to keep the version requirements
together with the ASSUME_PROVIDED variable. The syntax is rather
oelitish, e.g.

ASSUME_PROVIDED_MIN += "native:git_1.8.4 native:tar_1.2.3"