oelite/function.py: introduce start() and wait() methods
In order to support asynchronous execution of OEliteFunctions (on
which tasks are built), we need to split the run() method in two: one
that starts the execution, and one that waits for it to finish. Thus,
introduce start() and wait(), and make run() into a wrapper which
simply does start+wait, preserving the current behaviour.

The wait() method takes a boolean flag, indicating whether we're just
polling for whether the function completed (similarly to the WNOHANG
flag to the waitpid syscall) or if we want to synchronously wait for
it to actually finish (the latter is the default). Subclasses of
OEliteFunction which implement wait() and honour the poll argument are
supposed to return None in case the function hasn't finished yet.

The default start method does everything the old default run() did,
except that the return value is stashed in .result rather than
returned.

If a subclass wishes to override the default workdir and umask dance
it should implement its own start() method.

If a subclass doesn't implement wait(), its start() method must
populate the .result instance variable with either True or False. [It
does not make sense to override wait() but not start()].

To make it simpler for a subclass to override start() but still make
sure any subprocesses spawned have appropriate cwd and umask settings
while also ensuring that upon return to the caller, those are reset to
their original value, it can simply instead override _start().