oelite/task.py, oelite/baker.py: refactor stdio saving/restoring
When switching task functions to run asynchronously, we're going to
need a way to easily make sure all pre- and post-functions run with
the appropriate context. To this end, let OEliteBaker provide an easy
utility for any called function to easily restore its stdio file
descriptors.

This is of course rather ugly, as it relies on OEliteTask::run only
being called from a single place. We could also implement the
functionality in oelite/util.py, with backup_stdio() returning an
cookie which is supposed to be passed to restore_stdio() - the latter
would then both dup2 the copies back in place and close() the copies.