oelite/task.py: split run() method into several small helpers
As yet another preparation for parallel task execution, split the
setup and teardown parts of the run() method of OEliteTask into a few
helpers. When we then further split run() into start() and wait()
methods, the {save,apply,restore}_context helpers can be reused from
both of these. Even without the goal of parallel task execution,
breaking the code up into these helper functions makes it easier to
understand.

Note that this disables the tee'ing of output from the task to our
stdout when the debug flag is given - that's not compatible with
running tasks in parallel, so it would have to be removed at some
point. If people have strong feelings about this, I might consider
reinstating it in some form (possibly by forcing serial execution of
tasks).

There's still things one could do to optimize the fd handling, but
that's better suited for a few helpers in util.py. Concretely, this is
what I have in mind:

- There's no reason for every running task to have a file descriptor
  for /dev/null open for reading. We could make a simple utility
  function devnull, taking a single argument (the mode string) and
  either returning an existing open file instance corresponding to
  that mode, or opening /dev/null, caching the result, and returning
  it.

- Similarly, there's no reason for every task to have copies of the
  main process' stdio file descriptors. We could factor the
  saving/restoring facility into a small utility class: The
  constructor by default simply saves copies of std{in,out,err}, but
  one can pass an existing instance, in which case no extra file
  descriptors are created, and we also take note that we shouldn't
  close the copies when restoring. This of course relies on the
  "parent" object outliving all its descendants.