oelite/runq.py: refactor get_runabletask
In order to start tasks in parallel, we need a way to get a whole list
of runable tasks from baker.py. To that end, we're going to just steal
the cache of ready tasks from the runq. First, factor out the part of
get_runabletask which updates that cache.