oelite/function.py: allow ShellFunction to run asynchronously
This is more or less the last piece we need to run tasks in
parallel. We make use of the parent class doung the umask and cwd
dance for us by implementing _start rather than start. If we detect a
problem already at process creation time, we store False to .result,
which wait() can then pick up.

There's stuff to clean up, but this seems to work for now.