oelite/util.py: add stracehack
When stracing a python process, it can be difficult to know which
syscalls comes from which part of the code. A useful hack is to insert
easily recognizable no-op syscalls in various places, thus delineating
the actual syscalls made. As a bonus, by using a write to a certainly
invalid fd, we can also get some dynamic info dumped to the strace,
since the string also gets recorded.

This is obviously just meant for temporary debugging and profiling,
but its nice to have the utility available, hence this commit.