oelite/compat.py: add file for compat hacks
I'm going to set CLOEXEC on most file descriptors (mostly for
correctness - there's no reason for X:do_configure to inherit an open
fd for Y:do_compile.log). Annoyingly, python 2.7 doesn't expose the
necessary constants through the os or fcntl modules, despite them
being available on Linux since forever. It's not that an extra
fcntl(F_SETFD) takes that much time [and the atomicity is irrelevant
for our case], but it really annoys my inner optimization nerd. I also
have a few other ideas for things I might want to exploit but which
Python doesn't expose, so it's nice to have a place to put that.

I've made a very simple self-test, so one should be able to see if it
works by just running the compat.py file itself, maybe strace'ing it
as well to see which syscalls are actually made.