oelite/task.py: split run() into start() and wait()
Running a task is not simply a matter of calling .run on the
underlying OEliteFunction. Apart from the stdio setup, there's also
the concept of pre- and postfuncs that needs to be handled.

We don't want to complicate our caller more than necessary - he should
just have to split his call of task.run into task.start() and one or
more task.wait() calls. The natural thing to do is to run all the
pre-funcs in start(), and, if the function itself is succesful, run
the post-funcs in wait().

So in start(), we prepare the context (stdio fds) which all
OEliteFunctions should run in, apply it, and then, guarded by a
try-finally block, run the prefuncs and finish by starting the actual
task function. Should any of the prefuncs fail, we store a value to
.result which our wait-method can then pick up.

In wait(), we first check whether any of the prefuncs failed and if
so, return that information to the caller. Otherwise, we again apply
the appropriate stdio context, call the wait() method of the actual
task function, and provided that indicates succesful completion,
proceed to running the postfuncs.

Note that, by design, start() is not supposed to return a value. It
may throw an exception, or it may store a piece of state indicating to
wait() that it failed. The idea is to simplify callers, so they only
have to check one return value, and they don't care how the function
went until they explicit ask for the status by calling wait().

Just as its oelite/function.py counterpart, wait() is guaranteed to
return a True/False result if poll=False.