oelite/baker.py: record and store a few per-task-type timing statistics
Even if the times we record are not completely exact, knowing which
tasks typically take the longest can help guide optimization
efforts. Being put in tmp/profiling/, they're sufficiently out of the
way that I don't think a command-line option for suppressing them is
worthwhile.

Originally, this just displayed the average, but that by itself
doesn't say much - knowing more about the distribution can be quite
helpful, so also print the quartile values (three of which are also
known as min, max and median).