bb/utils.py: remove useless "safety" mechanism
Creating the file under a fake name for "safety" has a few drawbacks:

* if something does go wrong, we might leave the #new file behind

* we don't check whether the #new name actually already exists
  (unlikely, but still)

* it costs us an extra rename(2) system call, which appears to cost
  about .5 ms on average (!).

So just use the final name directly, and in the exceptional case, unlink
it - the remove() helper has "rm -f" semantics, so ENOENT is caught and
ignored.

The technique of creating a file under something other than its final
name is only useful if another process might concurrently be accessing
the to-be-created file (and hence might see a file in an inconsistent
state), but that's not relevant here.