TMP: profiling: add trace mechanism
It may sometimes be rather hard to follow the code flow and knowing how
many times a line of code get executed. Sticking in
oelite.profiling.trace() calls here and there is a slightly enhanced
version of "printf-debugging": The output doesn't get mixed up with
ordinary stdout, it gets stored permanently for later analysis, and we
also get small backtrace.