oelite/function.py: initialize .result member of PythonFunction to False
If an exception is thrown from somewhere and it trickles all the way up
to the baker's baking loop, we end up in the finally clause, calling
oven.wait_all(False) to retrieve the exit status for all already started
tasks. However, if we were running a PythonFunction, the "self.result =
self()" assignment isn't done, which means that the generic wait()
method ends up accessing a non-existing member of the PythonFunction
object. This causes another exception to be thrown (AttributeError:
'PythonFunction' object has no attribute 'result'), and that is what
ends up getting printed to the terminal, effectively masking the
underlying problem.

Fix this by setting .result of a PythonFunction to False initially -
that way, the task is (correctly) registered as having failed, and the
exception text gets written to the terminal.