oelite/function.py: open-code util.shcmd in ShellFunction class
As preparation for allowing us to run tasks in parallel, take the meat
of the util.shcmd helper routine, specialize to the (mostly default)
values we pass, and make it a helper function in the ShellFunction
class instead. This will allow us to do some further surgery,
splitting it into start() and wait() methods, which is obviously
necessary when we want to avoid running the whole task function
synchronously.