oelite/baker.py: move most baking logic inside oven
I need an abstraction for keeping the set of currently running tasks,
and "run queue" is already taken. Instead, let's take the kitchen
vocabulary a little further, and have the baker use an oven to do the
cooking. It's a really smart oven, as it knows how to check if a dish
is done, if it went well and what to do in either case.

For now, the baker hasn't learned how to make use of the oven's
multitasking features, so he still bakes one thing at a time.

I'm aware that oven.py is a little import-happy, but I'd rather trim
that later.