oelite/oven.py: tell what we're waiting for
It's nice to always be told what the process is doing, so when there's
a chance that we're not otherwise going to produce output for a while,
print out a list of the currently running tasks. We deliberatly wait
until a few iterations have gone by, so that we don't print anything
if a task was already finished or finished soon enough.