oelite/task.py: move function assignment to task::prepare
If we want to do something intelligent based on knowing if a function
is asynchronous or not, the OEliteFunction instance needs to be
available before we put the task in the oven. So move the
meta.get_function() call to task::prepare, and call that as soon as we
put the task on the pending list.