oelite/baker.py: record and store detailed task timings
All pre- and post-funcs run synchronously in the main oe process. Some
of these tend to be rather time-consuming, so it would be better if
they could be made into tasks of their own. Otherwise, the oven risks
being underused for a long time. Suppose some do_install script
finishes, and we go on to execute the associated postfuncs, which may
take > 15 seconds to complete. Meanwhile, some do_compile task has
finished, but there's no way we'll notice before we're done with those
postfuncs.

Knowing which tasks' pre- and postfuncs take a long time is useful for
finding recipes (or maybe whole classes) that are worth optimizing.