oelite/baker.py: manage the oven more cleverly
... or at least I thought so. My idea was the by knowing whether a
task is going to run asynchronously or not (and if so, whether it is a
do_compile or something else), one could schedule the most appropriate
task in order not to overload the system and cause excessive context
switches. However, this actually sems to make things take a longer
time, so this needs more thought. I'm still making this commit and
then reverting it so this gets recorded somewhere.