classes/core.oeclass: spawn fewer children
The absolutely most expensive syscall we use is clone() - when doing
an "oe bake world", it seems to take more than 0.1 seconds to
complete(!). I'm guessing that's because of our completely insane
memory consumption, so that the kernel has lots of vmas to mark COW.

One of the places we fork a lot is in do_install - in fact, that task
is responsible for 5928 out of 14917 clone calls during an oe bake
world. The reason is the we execute three shell commands for every
single library or binary. Let's instead just spawn a single shell
instance which will do the three commands.

We apparently want to do all three commands even if one fails (that's
what the current code does, anyway), so I've kept that behaviour for
now. I didn't find a neat way to instruct the shell to "run all these
commands, exit with failure if one failed", so I do that manually. I
threw in set -x so that one can match up error output with the command
generating it. That also eliminates the need for printing the commands
we're doing.

Since stdout and stderr is already pointing at do_install log file,
there's really no reason to do a lot of work to redirect those to
pipes we can read from, only to spit it out to the same log file in
case something went wrong.

While here, move the pathprefix "computation" and log printing below
our "nothing to do anyway" tests, and define the *cmd variables in the
order they are run.

Finally, commands is long deprecated, so use subprocess instead.