classes/package.oeclass: do fewer stat calls during do_split
We do 100000s of stat system calls, and while the linux kernel's
dcache makes that reasonably inexpensive, it's still a lot of useless
context switching and path walking.

This implements a simple stat result cacher, which can be used for the
most common queries (isdir, islink, isfile, exists, lexists), as well
as receiving the stat result itself (False if the file didn't
exist). It is obviously only valid as long as the user knows that the
caller doesn't do anything that might invalidate it, but that's
relatively easy to verify in the tight inner loops where this will be
most useful.