bb/utils.py: cosmetic improvements of copyfile
None of grand total of 1 callers pass the newmtime parameter, so kill
that. The return values are rather heterogenous (can be a stat result,
an mtime, or False). Return True of False, eliminating an lstat system
call when creating symlinks. In about 100% of all cases, the source
file is already readable by us, so eliminate the two redundant chmods.