classes/stage.oeclass: make utility function for extracting tarfile
The tarfile python module is... not written with performance in mind,
to say the least. In order to experiment with alternate ways of
achieving the extraction, create a simple oelite.util function to do
this - that way, we can replace that code without affecting the
metadata hash of each and every package.