core.oeclass: delete dead code
The parser insists that variable names get expanded when we do an
override assignment, so this code is never reached - inserting debug
statements and doing an oe bake world confirms that. Removing this dead
code allows us to make the implementation of the override mechanism
completely private to meta/dict.py.