oelite/parse/oeparse.py: intern function bodies
We store the text of all the python and bash functions defined in
recipes/classes verbatim. For functions defined in common classes, that
ends up duplicating the same strings needlessly. For example, the text
for do_packageqa alone amounts to 25KB * 222 recipes ~ 5.3MiB.

This only affects the cold-cache case, since we do the
force-everything-through-deepcopy-with-forced-string-interning dance
when reading in cached recipes.