oelayout.conf: remove OVERRIDES from FILESPATH
It's a little-used but somewhat expensive feature to have oe-lite search
for files in recipedir/X/Y for every value of Y in OVERRIDES (and X
running through the three well-known and widely-used ${P}, ${PN} and
files). The one recipe where this was used was easily modified to avoid
relying on this.