oelite/runq.py: temporarily do double bookkeeping in _[gs]et_provider
The runq.provider table is used to map a triple (type, name, version) to
a package (as in, an OElitePackage). Python has an excellent data
structure readily available for that: dict. Since None is just as
hashable as strings and work equally well inside tuples, this also ends
up simplifying the code.

As preparation for getting rid of the runq.provider table completely,
implement its replacement in parallel. This serves two purposes:

(a) check that the two implementations are equivalent

(b) measure the cost of the new implementation (roughly: how much does
"building dependency tree" time increase)