oelite/dbutil.py: only use (file, line) for collecting stats
In some places (e.g. cookbook.py::get_packages), we build the query
string dynamically, which means we get lots of single-entry stats. It's
more useful to collect these, so don't use the query string as part of
the key. But it's still nice to have part of the query printed, so stash
a sample the first time we get called from a specific line.