oelite/task.py: fix confusing regexp
First of all, a non-greedy ".*?" with no zero-width assertions
or anything else following is always guaranteed to match the empty
string, so it contributes nothing.

The once-compiled TASKFUNC_RE is only used in a single place, the
.meta() method. There, we only extract match_group(0), which is the
entire string matched, so there's no reason for the regexp to have a
capturing parentheses. It was probably there due to some confusion as to
whether the strings in the "emit" list we're building should contain the
"do_" prefix or not. It should, as also witnessed by the

        if not emit_task.startswith("do_"):
            emit_task = "do_" + emit_task

and this is indeed also what the current regexp and its use
achieves. This just makes the code a little less misleading.