lib/oelite: flush stdout/stderr before replacing underlying fd
oebakery contains an ancient commit (5b2740b70, Force stdout and stderr
into un-buffered mode). I'm pretty certain the source of the FUBAR
experienced is caused by replacing the file descriptor behind the
stream's back - if there's stuff left in the buffer when we do the dup2,
that will end up getting flushed to whatever file the stream points to
when the buffer is full the next time around.

It is probably mostly a problem when stdout is in full-buffered mode,
since we usually print whole lines, so the problem isn't as apparent in
a terminal (but one can easily reproduce it by just piping to cat).

Having stderr unbuffered makes perfect sense, but to be on the safe side
I also flush that in apply_context and its counterpart.

To reproduce the problem, and see that this fixes it, I've had this
chunk sitting near the top of the __init__ function in baker.py. Why
it's not just 'sys.stdout = os.fdopen(sys.stdout.fileno(), "w")' is left
as a little puzzle for the reader.

        if (not hasattr(oebakery, "__version__") or
            int(oebakery.__version__.split(".")[0]) < 5):
            origfd = sys.stdout.fileno()
            sys.stdout.flush()
            fd = os.dup(origfd)
            sys.stdout.close()
            os.dup2(fd, origfd)
            os.close(fd)
            sys.stdout = os.fdopen(origfd, "w")