oelite/pyexec.py: burn after running
There's no reason to keep the metadata about which hook functions have
been run around after the "event" (mid_recipe_parse etc.) have passed -
that just wastes memory (and time) by copying lots of lists within dicts
within dicts to thousands of task metadata instances. This is also in
line with add_hook (weakly) setting the emit flag to "" for the hook
functions.

The 'name' parameter to exechooks gets renamed to 'hookname', since name
is also used for a different purpose inside the loop, and we need to
original value after the loop.