oelite/runq.py: prepare for simplifying get_recipe_provider
There's really no reason to do a complicated

  get_recipe ->
  get_recipes ->
  translate OElitePackage to integer id ->
  lookup that id in the package table ->
  translate the recipe id to an OEliteRecipe

when an OElitePackage has a pointer to its defining recipe. There are
lots of places where we're not afraid to dig around the internals of the
various OEliteFoo objects, so upholding this particular maze of
indirection is just silly.

Just to be sure, this just checks that the alternative and obvious
implementation matches the current.