minimal.conf: assume git >= 1.8.4
This is partly to introduce an example of ASSUME_PROVIDED_MIN. But this
is not just some random git version - I've had a difficult-to-debug
problem which wouldn't have happened with a new enough host git.

The problem was a SRC_URI specifying a git repository accessed via a
non-standard ssh port, causing the ingredients subdir to contain a
colon. Subsequent cloning from that dir (during do_unpack) was,
pre-1.8.4, interpreted as an scp-style url rather than a local path,
which obviously breaks. Commit 60003340cda05f5ecd79ee8522b21eda038b994b
in git.git handles this case much more sanely.