oelite/oven.py: periodically print overall status
I was asked to add a little information about the number of pending
tasks etc. For now, this is the best I could come up with.

The numbers which I could think of that might be interesting:

  time-since-start
  #baking, #succeeded, #fails
  #pending, #remaining, #total

Of these, only #pending depends on stuff the oven doesn't
directly have access to, so we have the baker pass it a callable it can
use to find out.