oelite/meta: faster deep-copying
The deepcopy function in Python's standard library is extremely
slow. I've hacked together a C implementation of parts of copy.deepcopy,
which I'm in the process of upstreaming to CPython - but it will never
get applied/backported to 2.7, so here's a version we can use.

This cuts 60% from the deepcopying time, which translates to 25% of the
metadata hash computation time. I'm somewhat surprised the gain is not
larger than that, but we do have a few somewhat large/deep structures
hidden in e.g. the __overrides and __mtimes "variables".

For now, it does require an explicit action from the user, and having
Python headers available: Go to lib/oelite/meta and execute

  gcc -g -O2 -o _copy.so -fPIC -shared -I/usr/include/python2.7 _copy.c

I don't know if there's a good way of doing this automatically or
advising the user to do it - the try..except does ensure that there's no
ill effect of not reading meta/core commit logs carefully.