oelite/meta: faster deep-copying
The deepcopy function in Python's standard library is extremely
slow. I've hacked together a C implementation of parts of copy.deepcopy,
which I'm in the process of upstreaming to CPython - but it will never
get applied/backported to 2.7, so here's a version we can use.

This cuts 60% from the deepcopying time, which translates to 25% of the
metadata hash computation time. I'm somewhat surprised the gain is not
larger than that, but we do have a few somewhat large/deep structures
hidden in e.g. the __overrides and __mtimes "variables".

We use the new oelite.cext infrastructure to compile it automatically
the first time we try to import it (and again if the source file should
change).

With this, there's no longer any need to hook into
copy._deepcopy_dispatch, since the interning is done by the C module.