oelite/meta: faster deep-copying
The deepcopy function in Python's standard library is extremely
slow. I've hacked together a C implementation of parts of copy.deepcopy,
which I'm going to try to upstream to CPython - but for now, we might as
well benefit.

It does require an explicit action from the user: Go to lib/oelite/meta
and execute

  gcc -g -O2 -o _copy.so -fPIC -shared -I/usr/include/python2.7 _copy.c

This cuts 60% from the deepcopying time, which translates to 25% of the
metadata hash computation time. I'm somewhat surprised the gain is not
larger than that, but we do have a few somewhat large/deep structures
hidden in e.g. the __overrides and __mtimes "variables".