oelite/function.py: implement _start and wait in PythonFunction
For now this is just a copy of the generic implementations in
OEliteFunction.

When we implement async python functions in terms of fork(), we could
just do the umask and chdir in the child, but since we only do some
python functions asynchronously, we'd have to duplicate the
try..finally stuff in the synchronous case, and that's not worth it
for saving four system calls in the parent.