oelite/function.py: allow python functions to set __async flag
Well, sort of. For now we just implement retrieving
e.g. do_fetch[__async], but it has to be False.

I do get_flag with expand=CLEAN_EXPANSION to allow me to do something like

do_fetch[__async] = "${__ASYNC_FETCH}"

with the latter set in local.conf, or not set at all. The bool(int(
... or 0)) dance is so that all of unset, "", "0", "1", False, True
etc. do the expected thing. The double underscores are to ensure that
these variables and flags to not affect metadata hashes.