oelite/function.py: implement support for asynchronous PythonFunction
The list of PythonFunction tasks includes at least:

  do_fetch
  do_unpack
  do_stage
  do_split
  do_package

These are all rather I/O bound, and do_fetch in particular is prone to
stall the entire build if it is trying to fetch from an unresponsive
or just slow server - worst case, with our default timeout settings,
we can end up waiting 10 minutes, during which we may completely fail
to start other tasks.

This implements support for making a particular python task function
run asynchronously; simply add e.g.

do_fetch[__async] = True

above the do_fetch definition in fetch.oeclass.

In order for a task function implemented as a PythonFunction to safely
run asynchronously, it must not rely on mutating state in the OE-lite
process. Checking that is a rather tedious and error-prone job, so
this is mostly an experimental feature for now.