oelite/lockfile.py: rough lockfile implementation
If we want to do some mutual exclusion in the ingredients dir or
elsewhere, we will probably need some inter-process mechanism. Locking
a file using flock() is probably the simplest, and has the important
property that the necessary OS handles are actually exposed in Python
2.7.

I've tried to bolt a timeout mechanism on top, which we most likely
want. It's all entirely untested.