package.oeclass: make do_split more robust
Inside the loop, we do a lot of "files += foo". But mutating a list
being iterated over isn't guaranteed to work, and is in fact likely to
blow up if we ever start using Py3. So do it this way instead. (This
does mean we iterate over the files in essentially the opposite order
of what we used to, but that shouldn't matter, since this is still
per-package, and we do not change the directory tree under ${D} during
the iteration).