lib/oelite/oven.py: remember when tasks were added to and removed from oven
I have some postprocessing in mind which can help visualize which
tasks were running concurrently. For that we need to hang on to the
time stamps for when a task was put into the oven and removed
again. So record those using attributes of the OEliteTask instead of
the ephemeral .starttime dict. Incidentally, this should also address
some valid points kimrhh raised about the semantics of the remove()
method; now .add and .remove are completely symmetric.