lockfile.py: try to avoid (hack around) a race condition
The bottom-most changes in this patch are mostly cosmetic; semantically,
the only thing they do is an extra initial check of expired before
entering the loop for the first time. However, that should help
highlight a problem: If the signal fires while we're not blocked in
flock(), e.g. between evaluating the loop condition and calling flock(),
we'll never receive a signal that will bring us out of flock(), making
us block forever.

This is really just an instance of the prototypical race condition

  check condition
                                    <- condition changes
  if ok, do a blocking system call

We can work around this by having the signal fire every .1 seconds after
the timeout expires. We may still enter flock() after the timeout has
actually expired, which may of course then succeed, but we should never
block indefinitely.

Instead of .1 second, one may want to use something relative to the
given timeout (e.g. 1/100th of that).

----

My initial idea for solving this is so bad that it deserves being
written down: Just raise the ETIMEDOUT exception from the signal
handler.

You see why that's a terrible idea? No? The signal may fire after
flock() has succeeded, but before Python has gotten around to calling
the __exit__ method (thus disabling the timer). In that case, we'd own a
lock but lie to our caller that we didn't get it in time.