oelite/task.py: let oelite.log handle fd redirections
The StdioSaver was meant to eventually become shared between tasks, but
it's really much simpler just saving a copy of each of stdio fd once at
at startup as its done in oelite.log. Moreover, letting the logging
framework know the context allows it to, for example, write critical
messages to the terminal preceded by the current task name if we're in
task context.