oelite/util.py: avoid races in makedirs
The stdlib utility function os.makedirs raises EEXIST if the leaf
directory already exists. Our makedirs helper tries to avoid that by
checking for that condition upfront.

However, when doing tasks in parallel, two different processes may call
makedirs concurrently; both may initially see that the directory does
not exist, one will successfully call os.makedirs(), while the other
will get an EEXIST exception. Debugging that would be no fun, since the
code clearly checks whether the path exists... This reimplements the
helper to avoid such races.

The previous implementation would also silently return success in the
case where 'path' exists and is a regular file, which would in all
likelyhood just lead to ENOTDIR later when the caller tries to create
'path + "/foo"'. It's better to fail early, and this is also consistent
with 'mkdir -p' failing if the target exists but is not a directory.