add "oe test" subcommand
I've tried to add small unit tests here and there, especially when
low-level OS stuff is involved. This is an attempt at creating a place
to put all these, allowing them to be run without weird and ad hoc
command lines.

The first test case I'm adding is a bit complex, but it's the simplest I
could come up with to provoke the race in our makedirs helper. That race
may show its ugly face when we do stuff in parallel processes and be
very hard to debug ("what do you mean EEXIST, the code obviously tests
for that before calling os.makedirs!"). Before fixing any race one needs
a test case that reliably reproduces it.