oelite/util.py: introduce hash_file and chunk_iter helpers
We get the sha1 or md5 digest of files in a couple of places by doing

  m = hashlib.md5()
  with open("foo") as f:
    m.update(f.read())
  h = m.hexdigest()

But that .read() call can be rather expensive if the file in question is
a 100M tarball (opencv, I'm looking at you); it temporarily bumps our
memory usage quite a lot, and it's also terribly inefficient (lots of
page faults and completely blowing the cpu cache and TLB first by
populating the gigantic buffer, then again when reading through it).

So create a generator which simply yields a given file in chunks. Then
the above would be:

  m = hashlib.md5()
  for chunk in iter_chunks("foo"):
    m.update(chunk)
  h = m.hexdigest()

but since we do that middle part a lot, create another helper that does
that and returns the passed-in hasher (we sometimes feed more than just
one file to the same hasher, so we don't finish up by calling
.hexdigest). Then the above could be simply

  h = oelite.util.hash_file(hashlib.md5(), "foo").hexdigest()