oelite/fetch/sigfile.py: make it a little more robust
I'm not sure this is really worth the complexity it adds, but it
annoyed the correctness nerd in me.

If

- a .oe file has several recipe types,
- we build at least two of them,
- they happen to have a different set of SRC_URIs, and
- we lack signatures for those

each do_fetch will write a .sig file, but only containing the
signatures pertaining to the SRC_URIs for that variant. The end result
is a .sig file not containing all the signatures it should, and thus
the next build is again going to fail with a "signatures changed,
rebuild needed".

We can avoid that by updating our view of the world before writing the
new contents.

Doing do_fetch in separate processes would make this a slightly bigger
problem than just having to run "oe bake" n+1 times; without
synchronization, we could end up generating corrupt .sig files. So
while we're at it, prepare for that by throwing in flock(LOCK_EX).