oelite/meta/dict.py: share (value, dependency set) tuples
The values we store in the .expand_caches are immutable (if we make a
change to one of the vars in the dependency set, the entry gets
deleted). This means we may share those tuples among all the metadata
instances, saving a significant amount of memory.

First, this requires changing the second member to a frozenset, and
that reveals a (harmless) bug in the override handling; one shouldn't
mutate the dependency set one gets from a _get call (the effect was
just to add "OVERRIDES" to the dependency set for the "OVERRIDES"
variable itself). If we have other such bugs hiding somewhere, the
change to using a frozenset should flush them out as well.

Then it's just a matter of creating two dicts for holding the interned
frozensets and tuples, using the same strategy as Python uses
internally for interning strings. As a bonus, we get to drop the
optimization where we used None instead of an empty set, since we now
use the singleton empty frozenset.

The saving is around 7% (40M for oe bake world), with no measurable
effect on initialization time. I've done the usual dump of the
metadata before and after to check that this does not affect hashes.

The dict holding the interned tuples ends up growing to a few MB, so
clear the caches after we've done all the hash computations. That may
or may not reduce our memory footprint by those MB, depending on the
details of Python's and glibc's memory allocation strategy.