cookbook.py: use intern as text_factory
I was wondering how we ended up creating ~5000 copies of the strings
"machine" and "arm-926ejs-linux-gnueabi" (each thus costing about 3M
of memory). It turns out these are the .type and .arch members of the
OElitePackage instances, and the reason they each have their own copy
is that the string objects are fresh out of the sqlite database.

Let's ask pysqlite to intern every string it passes back to us. This
saves a lot of memory, and is virtually free.