oelite/meta/meta.py: make ExpansionStack smaller
The memory footprint of an ExpansionStack instance is 72 bytes for the
instance itself, 280 bytes for the __dict__ attribute, and then the
memory for the actual list stored at the .stack attribute. We can get
rid of the first two contributions by simply letting an ExpansionStack
be a list with one extra method that does the sanity checking.

Conveniently, the builtin list type even already has a .pop method
which does the right thing. Also, __init__ and __len__ are already
taken care of. Nobody calls __str__ directly allowing them to pass the
optional argument and str(foo, "something else") would be a runtime
error, so just make __str__ have the usual prototype.

The .python attribute never seems to have been used for anything.

Saving 352 bytes per instance adds up to about 1.5M for an oe bake
world.